Avoid os_log decoding errors in perception check
We currently pretty-print a diff format using a figure space, which is
in the range of unicode that results in an error being printed to the
console instead of the log message. While the message is a little less
readable without the diff, removing this formatting fixes the error.

Fixes #148.